int => long

+++ rapid-common-util/src/main/java/com/github/rapid/common/util/page/PageQuery.java
@@ -7,16 +7,16 @@ package com.github.rapid.common.util.page;
 public class PageQuery implements java.io.Serializable{
     private static final long serialVersionUID = -8000900575354501298L;
     
-    public static final int DEFAULT_PAGE_SIZE = 10;
+    public static final long DEFAULT_PAGE_SIZE = 10;
     /** 页数  */
-    private int    page;
+    private long    page;
     /** 分页大小 */
-    private int    pageSize = DEFAULT_PAGE_SIZE;
+    private long    pageSize = DEFAULT_PAGE_SIZE;
 
     public PageQuery() {
     }
 
-    public PageQuery(int pageSize) {
+    public PageQuery(long pageSize) {
         this.pageSize = pageSize;
     }
     
@@ -25,33 +25,33 @@ public class PageQuery implements java.io.Serializable{
         this.pageSize = query.pageSize;
     }
     
-    public PageQuery(int page, int pageSize) {
+    public PageQuery(long page, long pageSize) {
         this.page = page;
         this.pageSize = pageSize;
     }
 
-    public int getPage() {
+    public long getPage() {
         return page;
     }
 
-    public void setPage(int page) {
+    public void setPage(long page) {
         this.page = page;
     }
     
-    public PageQuery page(int page) {
+    public PageQuery page(long page) {
     	this.page = page;
     	return this;
     }
 
-    public int getPageSize() {
+    public long getPageSize() {
         return pageSize;
     }
 
-    public void setPageSize(int pageSize) {
+    public void setPageSize(long pageSize) {
         this.pageSize = pageSize;
     }
     
-    public PageQuery pageSize(int pageSize) {
+    public PageQuery pageSize(long pageSize) {
     	this.pageSize = pageSize;
     	return this;
     }
